chess/requirements.txt
autogen-agentchat==0.5.7
autogen-ext==0.5.7
python-chess
chess-board
rich
python-dotenv
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry.exporter.otlp.proto.grpc==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-flask==0.54b1
opentelemetry-instrumentation-wsgi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
opentelemetry-instrumentation-httpx==0.54b1
opentelemetry-instrumentation-openai==0.40.6
opentelemetry-instrumentation-requests==0.54b1
opentelemetry-instrumentation-mcp==0.40.7